Module Equinoxe

Equinoxe library interface.

module type API = sig ... end

It offers OCaml types to manipulate the Equinix API.

Build your own API

module type Backend = sig ... end

This module gathers all the methods you need to be able to execute HTTP requests to contact an API server. It must send application/json request.

module Make (B : Backend) : API with type 'a io = 'a B.io

Factory to build a system to communicate with Equinix API in a strongly-typed way using the Backend gathering system.