Module Equinoxe.Json

Utilities for Equinoxe-cli.

The Json module provides helpers to manipulate JSON objects.

type t

Abstract type to represent JSON objects.

val error : string -> t

error msg produces a type t from an error message.

val of_string : string -> t

of_string str takes a string str representing a JSON and transform it into an t object you can manipulate with this module.

val geto : t -> string -> t

geto json field_name returns the t value associated to the field_name.

val geta : t -> int -> t

geta json nth. returns the t value associated to the nth element in the json.

val to_int_r : t -> (int, string) Stdlib.result

to_int_r json transforms the json object into an int result with a printable error in case of failure.

val to_string_r : t -> (string, string) Stdlib.result

to_string json transforms the json into a string result with a printable error in case of failure.

module Infix : sig ... end