Module Json.Private

/!\ Private area should not be used while using the API!

val filter_error : t -> t

filter_error json checks for a possible errors fields in json. If it's not the case, the json is not modified.