Module Equinoxe.Make

Factory to build a system to communicate with Equinix API, using the S communication system.

It is the signature of the API of the website.

Parameters

module C : S

Signature

type t

Abstract type t represents the information known by the API system.

val create : endpoint:string -> ?token:[ `Default | `Str of string | `Path of string ] -> unit -> t

create opts returns an t object, you need to manipulate when executing requests.

module Auth : sig ... end

This module manages API part related to authentifications.

module Orga : sig ... end

This module manages API part related to organizations.

module Users : sig ... end

This module manages API part related to users.

module Metal : sig ... end