Module Make.Auth

This module manages API parts related to authentification.

val get_user_api_keys : t -> json io

get_user_api_keys t returns the keys available for the current user.

val post_user_api_keys : t -> ?read_only:bool -> description:string -> unit -> json io

post_user_api_keys t ~read_only ~description () creates a new API key on Equinix. Default value to read_only is true.

val delete_user_api_keys_id : t -> id:string -> unit -> json io

delete_user_api_keys_id t ~id () deletes the key referenced by id from the user keys.