Module Make.Devices

This module manages API parts related to devices.

type action =

| Power_on

| Power_off

| Reboot

| Reinstall

| Rescue

Actions executable with a device.

type os =

| Debian_9

| Debian_10

| NixOs_21_05

| Ubuntu_18_04

| Ubuntu_20_04

| Ubuntu_21_04

| FreeBSD_11_2

| Centos_8

Os available when creating a new device.

type location =

| Washington

| Dallas

| Silicon_valley

| Sao_paulo

| Amsterdam

| Frankfurt

| Singapore

| Sydney

Locations available when deploying a new device.

type plan =

| C3_small_x86

| C3_medium_x86

Server type when deploying a new device.

type config = {

hostname : string;

location : location;

plan : plan;

os : os;

}

This type represents the configuration wanted for a device.

val os_to_string : os -> string

os_to_string os converts an os into a string understandable by the API.

val location_to_string : location -> string

location_to_string facility convert a facility into a string understandable by the API.

val plan_to_string : plan -> string

plan_to_string plan convert a plan into a string understandable by the API.

val get_devices_id : t -> id:string -> unit -> Json.t

get_devices_id t ~id () returns a Json.t that contains information about the device specified by id.

val get_devices_id_events : t -> id:string -> unit -> Json.t

get_device_id_events t ~id () retrieves information about the device events.

val post_devices_id_actions : t -> id:string -> action:action -> unit -> Json.t

post_devices_id_actions t ~id ~action () executes an action on the device specified by its id.

val delete_devices_id : t -> id:string -> unit -> Json.t

delete_devices_id t ~id () deletes a device on Equinix and returns a Json.t with the result.