Module Make.Event

This module deals with events that occure in Equinix.

type id

Unique identifier.

val id_of_string : string -> id

id_of_string str returns a string referencing the event in Equinix.

val id_to_string : id -> string

id_to_string id returns a string representation of the id.

type t = {

id : id;

state : State.t;

event_type : string;

body : string;

created_at : Equinoxe__Equinoxe_intf.Date.t;

}

A representation of an event.

val t_of_json : Ezjsonm.value -> t

t_of_json json extracts information about event from json. If the parsing fails, it raises an Ezjsonm.Parse_error exception.

val to_string : t -> string

to_string t returns a string representation of t.

val pp : t -> unit

pp t prints in a human readable way the t value.