Module Make.Orga

This module manages API part related to organizations.

val get_organizations : t -> Json.t

get_organizations t returns all the organizations associated with the token.

val get_organizations_id : t -> id:string -> unit -> Json.t

get_organizations_id t ~id () returns the Json.t that is referenced by the id given in parameter.