Module Make.Project

This module manages API parts related to projects.

type id

Unique identifier for the Equinix API.

type config = {

id : id;

name : string;

created_at : Equinoxe__Equinoxe_intf.Date.t;

updated_at : Equinoxe__Equinoxe_intf.Date.t;

}

Representation of an Equinix Project.

val id_of_string : string -> id

id_of_string str converts str into an id.

val string_of_id : id -> string

string_of_id id returns the string that represents the id.

val to_string : config -> string

to_string config returns a string representation of the config.

val get_all : t -> config list io

get_all t returns all projects associated with the token.

val get_from : t -> id:id -> config io

get_from t ~id returns the config of the project that is referenced by the id given in parameter.

val pp : config -> unit

pp config prints a human readable Project config.