Module Make.Users

This module manages API part related to the user.

val get_own_id : t -> Json.t

get_own_user t returns informations about the user linked to the API key.

val get_api_keys : t -> Json.t

get_api__keys t returns the keys available for the current user.

val add_api_key : t -> ?read_only:bool -> string -> Json.t

add_api_key t ~read_only description creates a new API key on Equinix. Default value to read_only is true.

val del_api_key : t -> string -> Json.t

del_api_key t key_id deletes the key referenced by key_id from the user keys.