Module Equinoxe.MakeFriendly

Factory to build a system to communicate with Equinix API in a strongly-typed way using the Backend gathering system.

It offers OCaml types to manipulate the Equinix API.

Parameters

module B : Backend

Signature

type 'a io = 'a B.io

type t

Abstract type t represents the information known by the API system.

val create : ?address:string -> ?token:string -> unit -> t

create ~address ~token () returns an t object, you need to manipulate when executing requests. Default address is https://api.equinix.com/metal/v1/ and default token is empty.

module Orga : sig ... end

A module to interact with Equinix organization.