Module Utils.Reader

The Reader module provides helpers to read from files.

val read_token_opt : string -> string option

read_token_opt path reads the first line of the file path and returns it as Some token if it founds one, otherwise None. The token must be on the first line of the file.