Module Utils.Sys

The Sys module provides helpers for system calls. It's an extension of Sys module.

include module type of Stdlib.Sys

val argv : string array

val executable_name : string

val file_exists : string -> bool

val is_directory : string -> bool

val remove : string -> unit

val rename : string -> string -> unit

val getenv : string -> string

val getenv_opt : string -> string option

val command : string -> int

val time : unit -> float

val chdir : string -> unit

val mkdir : string -> int -> unit

val rmdir : string -> unit

val getcwd : unit -> string

val readdir : string -> string array

val interactive : bool Stdlib.ref

val os_type : string

type backend_type = Stdlib__Sys.backend_type =

| Native

| Bytecode

| Other of string

val backend_type : backend_type

val unix : bool

val win32 : bool

val cygwin : bool

val word_size : int

val int_size : int

val big_endian : bool

val max_string_length : int

val max_array_length : int

val max_floatarray_length : int

val runtime_variant : unit -> string

val runtime_parameters : unit -> string

type signal_behavior = Stdlib__Sys.signal_behavior =

| Signal_default

| Signal_ignore

| Signal_handle of int -> unit

val signal : int -> signal_behavior -> signal_behavior

val set_signal : int -> signal_behavior -> unit

val sigabrt : int

val sigalrm : int

val sigfpe : int

val sighup : int

val sigill : int

val sigint : int

val sigkill : int

val sigpipe : int

val sigquit : int

val sigsegv : int

val sigterm : int

val sigusr1 : int

val sigusr2 : int

val sigchld : int

val sigcont : int

val sigstop : int

val sigtstp : int

val sigttin : int

val sigttou : int

val sigvtalrm : int

val sigprof : int

val sigbus : int

val sigpoll : int

val sigsys : int

val sigtrap : int

val sigurg : int

val sigxcpu : int

val sigxfsz : int

exception Break

val catch_break : bool -> unit

val ocaml_version : string

val enable_runtime_warnings : bool -> unit

val runtime_warnings_enabled : unit -> bool

val opaque_identity : 'a -> 'a

module Immediate64 : sig ... end

val path_from_home_dir : string -> string

path_from_home_dir path returns a new path from the home directory. It can be seen as an expansion of "$HOME/path".

It raises Not_found if the HOME value is not found.