Module type Equinoxe.API

It offers OCaml types to manipulate the Equinix API.

type 'a io

type t

Abstract type t represents the information known by the API system.

val create : ?address:string -> ?token:string -> unit -> t

create ~address ~token () returns an t object, you need to manipulate when executing requests. Default address is https://api.equinix.com/metal/v1/ and default token is empty.

exception Unknown_value of string * string

This exception represents an error when parsing works but the value received is unknown.

module Orga : sig ... end

A module to interact with Equinix organization.

module User : sig ... end

A module to interact with Equinix user.

module Auth : sig ... end

This module manages API parts related to authentification.

module Ip : sig ... end

This module manages API parts related to ips.

module Project : sig ... end

This module manages API parts related to projects.

module State : sig ... end

This module represents the state of a Device in the Equinix API.

module Event : sig ... end

This module deals with events that occure in Equinix.

module Device : sig ... end

This module manages API parts related to devices.