Module Equinoxe

Equinoxe library interface.

Manipulate Results

module Json : sig ... end

The Json module provides helpers to manipulate JSON objects.

module type API = sig ... end

It is the signature that matches the API of the website.

Build your own API

module type Backend = sig ... end

The S module gathers all the methods you need to be able to execute HTTP requests to contact an API server. It must send application/json request.

module Make (B : Backend) : API

Factory to build a system to communicate with Equinix API, using the S communication system.